TJDMX: - Splitted off the actual DMX engine into a separate library (TJDMXEngine); TJDMXEngine is more or less portable (i.e. only depends on TJSharedUI on Windows; some devices are disabled when building on other platforms, but ArtNet, ESP and LanBox work).
TJNP:
- Added SendTo and more variants of Send in Socket class
- Added Socket::SetBroadcast/SetReuseAddress/SetBlocking
- WebServer now uses Bytes to identify item sizes

TJShared:
- Data now uses Bytes to indicate size (which is a 64 bit integer)

git-svn-id: https://svn.pixelspark.nl/svn/doip/trunk@664 984cb85d-a4ae-4555-b754-76bc93324474

+++ EP/EP/src/epdownload.cpp
@@ -29,7 +29,7 @@ void EPDownloadedDefinition::OnCreated() {
 void EPDownloadedDefinition::OnDownloadComplete(ref<DataWriter> cw) {
 	if(cw) {
 		TiXmlDocument doc;
-		std::string data((const char*)cw->GetBuffer(), 0, cw->GetSize());
+		std::string data((const char*)cw->GetBuffer(), 0, (unsigned int)cw->GetSize());
 		doc.Parse(data.c_str());
 		TiXmlElement* root = doc.FirstChildElement("endpoint");
 		if(root!=0) {
